requirements.txt
container-app-conf
python-telegram-bot==12.0.0b1
telegram-click
pattern
emoji
sqlalchemy
alembic
prometheus-client